lambda_handlers/authorizer/requirements.txt
fastapi>=0.110.0
PyJWT>=2.8.0
structlog>=24.1.0
uvicorn>=0.29.0
